more work on game state, getting everything working together

+++ Minecraft Client/Assets/_Project/Scripts/Protocol/Packets/Serverbound/HandshakePacket.cs
@@ -19,7 +19,7 @@ public class HandshakePacket : Packet
 			List<byte> builder = new List<byte>();
 			builder.AddRange(VarInt.GetBytes(ProtocolVersion));
 			builder.AddRange(PacketStructureUtility.GetBytes(Address));
-			builder.AddRange(BitConverter.GetBytes(Port).ReverseIfLittleEndian());
+			builder.AddRange(BitConverter.GetBytes((ushort)Port).ReverseIfLittleEndian());
 			builder.AddRange(VarInt.GetBytes((int)NextState));
 			return builder.ToArray();
 		}
@@ -31,6 +31,6 @@ public class HandshakePacket : Packet
 
 	public int ProtocolVersion { get; set; } = NetworkClient.PROTOCOL_VERSION;
 	public string Address { get; set; }
-	public ushort Port { get; set; }
+	public int Port { get; set; }
 	public NetworkClient.ProtocolState NextState { get; set; }
 }
